pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai_ignore_manager"
version = "0.1.0"
description = "A CrewAI tool to automate .gitignore and .dockerignore management"
authors = [
  { name="Antigravity", email="antigravity@example.com" },
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "crewai",
    "python-dotenv",
    "crewai-tools"
]

[project.scripts]
ignore-manager = "crewai_ignore_manager.cli:main"

[tool.setuptools.packages.find]
where = ["src"] 

